requirements.txt
opencv-python
SpeechRecognition
pydub
pyAudio
baidu-aip
erniebot
playsound==1.2.2
